+++ DrivingLicense.java
@@ -2,7 +2,7 @@ public class DrivingLicense {
 
     public static void main(String[] args) {
         //inicializar una variable (del tipo que veas conveniente) que contenga los puntos del carnet de conducir que empiece con 0 puntos.
-        int points = 0;
+        byte points = 0;
         System.out.println("Has sacado tu carnet de conducir, felicidades tienes 12 puntos");
         // Actualiza los puntos aquí
         points += 12;
